Modularize Cpu code Fixes #8
