+++ src/bin/main.rs
@@ -1,7 +1,7 @@
 use z80_undead::cpu::Cpu;
 
 fn main() {
-    let mut cpu = Cpu::new();
+    let _cpu = Cpu::new();
     println!("Z80 Emulator initialized");
     // Add more functionality here as we develop the emulator
 }
