+++ src/cpu/instructions.rs
@@ -1,0 +1,98 @@
+use super::*;
+
+impl Cpu {
+    pub fn add_a(&mut self, value: u8) {
+        let a = self.a;
+        let result = a.wrapping_add(value);
+
+        self.set_flag(FLAG_S, result & 0x80 != 0);
+        self.set_flag(FLAG_Z, result == 0);
+        self.set_flag(FLAG_H, (a & 0x0F) + (value & 0x0F) > 0x0F);
+        self.set_flag(FLAG_PV, (a ^ value ^ 0x80) & (a ^ result) & 0x80 != 0);
+        self.set_flag(FLAG_N, false);
+        self.set_flag(FLAG_C, (a as u16) + (value as u16) > 0xFF);
+
+        self.a = result;
+    }
+
+    pub fn sub_a(&mut self, value: u8) {
+        let a = self.a;
+        let result = a.wrapping_sub(value);
+
+        self.set_flag(FLAG_S, result & 0x80 != 0);
+        self.set_flag(FLAG_Z, result == 0);
+        self.set_flag(FLAG_H, (a & 0x0F) < (value & 0x0F));
+        self.set_flag(FLAG_PV, (a ^ value) & (a ^ result) & 0x80 != 0);
+        self.set_flag(FLAG_N, true);
+        self.set_flag(FLAG_C, a < value);
+
+        self.a = result;
+    }
+
+    pub fn and_a(&mut self, value: u8) {
+        self.a &= value;
+
+        self.set_flag(FLAG_S, self.a & 0x80 != 0);
+        self.set_flag(FLAG_Z, self.a == 0);
+        self.set_flag(FLAG_H, true);
+        self.set_flag(FLAG_PV, self.a.count_ones() % 2 == 0);
+        self.set_flag(FLAG_N, false);
+        self.set_flag(FLAG_C, false);
+    }
+
+    pub fn or_a(&mut self, value: u8) {
+        self.a |= value;
+
+        self.set_flag(FLAG_S, self.a & 0x80 != 0);
+        self.set_flag(FLAG_Z, self.a == 0);
+        self.set_flag(FLAG_H, false);
+        self.set_flag(FLAG_PV, self.a.count_ones() % 2 == 0);
+        self.set_flag(FLAG_N, false);
+        self.set_flag(FLAG_C, false);
+    }
+
+    pub fn xor_a(&mut self, value: u8) {
+        self.a ^= value;
+
+        self.set_flag(FLAG_S, self.a & 0x80 != 0);
+        self.set_flag(FLAG_Z, self.a == 0);
+        self.set_flag(FLAG_H, false);
+        self.set_flag(FLAG_PV, self.a.count_ones() % 2 == 0);
+        self.set_flag(FLAG_N, false);
+        self.set_flag(FLAG_C, false);
+    }
+
+    // Implement other instruction methods...
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_add_a() {
+        let mut cpu = Cpu::new();
+        cpu.a = 5;
+        cpu.add_a(3);
+        assert_eq!(cpu.a, 8);
+        assert!(!cpu.get_flag(FLAG_Z));
+        assert!(!cpu.get_flag(FLAG_S));
+        assert!(!cpu.get_flag(FLAG_C));
+        assert!(!cpu.get_flag(FLAG_PV));
+    }
+
+    #[test]
+    fn test_sub_a() {
+        let mut cpu = Cpu::new();
+        cpu.a = 10;
+        cpu.sub_a(3);
+        assert_eq!(cpu.a, 7);
+        assert!(!cpu.get_flag(FLAG_Z));
+        assert!(!cpu.get_flag(FLAG_S));
+        assert!(!cpu.get_flag(FLAG_C));
+        assert!(!cpu.get_flag(FLAG_PV));
+        assert!(cpu.get_flag(FLAG_N));
+    }
+
+    // Add more instruction-related tests...
+}
